app/pyproject.toml
[project]
name = "alloydflanagan-wag"
version = "0.1.0"
description = "Personal web site for A. Lloyd Flanagan"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=2.3.0",
    "django-htmx>=1.23.0",
    "django-redis>=5.4.0",
    "django-structlog>=9.1.1",
    "django-waffle>=4.2.0",
    "django-widget-tweaks>=1.5.0",
    "gunicorn>=23.0.0",
    "psycopg>=3.2.7",
    "structlog>=25.3.0",
    "wagtail<5.3",
    "whitenoise>=6.9.0",
]

[dependency-groups]
dev = [
    "coverage>=7.8.0",
    "django-coverage-plugin>=3.1.0",
    "django-debug-toolbar>=5.2.0",
    "django-stubs>=5.2.0",
    "djhtml>=3.0.7",
    "factory-boy>=3.3.3",
    "flake8-django>=1.4",
    "freezegun>=1.5.1",
    "ipython>=9.2.0",
    "isort>=6.0.1",
    "mypy>=1.15.0",
    "ruff>=0.11.8",
    "ruff-lsp>=0.0.62",
    "types-freezegun>=1.1.10",
    "types-python-dateutil>=2.9.0.20241206",
    "types-requests>=2.32.0.20250328",
    "vcrpy>=5.1.0",
]

[tool.isort]
line_length = 120
known_django = "django"
known_wagtail = "wagtail"
sections = ["FUTURE","STDLIB","DJANGO","WAGTAIL","THIRDPARTY","FIRSTPARTY","LOCALFOLDER"]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip = "migrations"

[tool.ruff]
line-length = 120
exclude = ["migrations"]
